% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mop.R
\name{asMOPString}
\alias{asMOPString}
\title{Parameters to MOP String}
\usage{
asMOPString(parameters)
}
\arguments{
\item{parameters}{A \code{"numeric"} vector containing the coefficients.}
}
\value{
A \code{"character"} string with a \code{'mop'} structure.
}
\description{
This function builds a string with the structure of a \code{'mop'} function.
}
\examples{
param <- c(1,2,3,4)
asMOPString(param)

param <- 3.4
asMOPString(param)
}

