% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mte.R
\name{asMTEString}
\alias{asMTEString}
\title{Converting MTEs to strings}
\usage{
asMTEString(parameters, num = 5)
}
\arguments{
\item{parameters}{A \code{"numeric"} vector containing the coefficients.}

\item{num}{A \code{"numeric"} value which contains the denominator of the coefficient
in the exponential.}
}
\value{
A \code{"character"} string with an \code{'mte'} structure.
}
\description{
This function builds a string with the structure of an \code{'mte'} function.
}
\examples{

param <- -5.8
asMTEString(param)

param <- c(5.2,0.3,-3,4)
asMTEString(param)
 
}
