% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean}
\alias{clean}
\title{Remove Objects from Memory}
\usage{
clean(envir = globalenv(), n = 2)
}
\arguments{
\item{envir}{The currently active environment; by default It is the gloval environment.}

\item{n}{Number of garbage collection repetitions; by default \code{n = 2}.}
}
\description{
Clean the memory. Delete all the objects in memory and a garbage collection
takes place.
}
\examples{
## Run to clean the environment
clean()
clean(n=2)
}
