% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoTBFClass.R
\name{coercion-motbf}
\alias{coercion-motbf}
\alias{as.character.motbf}
\alias{as.function.motbf}
\alias{as.character.jointmotbf}
\alias{as.function.jointmotbf}
\title{Coerce MOTBF Objects to Character or Function}
\usage{
\method{as.character}{motbf}(x, ...)

\method{as.function}{motbf}(x, ...)

\method{as.character}{jointmotbf}(x, ...)

\method{as.function}{jointmotbf}(x, ...)
}
\arguments{
\item{x}{An object of class \code{'motbf'} or \code{'jointmotbf'}.}

\item{...}{Further arguments passed to or from other methods. Not used currently.}
}
\value{
\itemize{
\item \code{as.character}: Returns a character string representing the mathematical expression of the object.
\item \code{as.function}: Returns an executable R function that accepts numeric arguments to evaluate the MoTBF expression.
}
}
\description{
Converts \code{'motbf'} and \code{'jointmotbf'} objects into character string expressions or executable R functions.
}
\examples{

## Example 1
X <- rchisq(5000, df = 3)
P <- univMoTBF(X, POTENTIAL_TYPE = "MOP"); P
as.function(P)(10)

## Example 2
data <- data.frame(X = rnorm(100), Y = rexp(100))
dim <- c(3,2)
P <- jointmotbf.fit(data, dimensions = dim)
density <- as.function(P)(data[,1], data[,2])
sum(log(density))

}
