% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidation.R
\name{confusionMatrix}
\alias{confusionMatrix}
\alias{confusionMatrix.motbf_fit_cv}
\title{Confusion Matrix}
\usage{
confusionMatrix(x, ...)

\method{confusionMatrix}{motbf_fit_cv}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object used to select the appropriate method, such as one of class \code{'motbf_fit_cv'}.}

\item{...}{Further arguments. Not used currently.}

\item{digits}{An integer indicating the number of decimal places to round the results.}
}
\value{
A matrix or table containing the confusion matrix.
}
\description{
Computes the confusion matrix for a given fitted model object.
}
