% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{dimensionFunction}
\alias{dimensionFunction}
\title{Dimension of MoTBFs}
\usage{
dimensionFunction(P)
}
\arguments{
\item{P}{An object of class \code{"motbf"} and subclass 'mop' or \code{"jointmotbf"}.}
}
\value{
Dimension of the function.
}
\description{
Get the dimension of \code{"motbf"} and \code{"jointmotbf"} densities.
This function is deprecated; use \code{"getMotbfDim"} instead.
}
\seealso{
\link{univMoTBF} and \link{jointMoTBF}
}
