% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint.R
\name{eval.motbf}
\alias{eval.motbf}
\title{Evaluation of MoTBFs}
\usage{
eval.motbf(P, values)
}
\arguments{
\item{P}{An object of class \code{"univmotbf"} or \code{"jointmotbf"}.}

\item{values}{A list with the name of the variables equal to the values to be evaluated.}
}
\value{
If all the variables in the equation are evaluated, then a \code{"numeric"} value
is returned. Otherwise, an \code{"univmotbf"} object or a \code{"jointmotbf"} object is returned.
}
\description{
Evaluates a univariate (\code{"mop"}, \code{"mte"}) or joint distribution (\code{"jointmotbf"}) at a specific point.
}
\examples{
#' ## 1. EXAMPLE
## Dataset with 2 variables
X <- data.frame(rnorm(100), rexp(100))

## Joint function
dim <- c(3,3) # dim <- c(5,4)
P <- jointmotbf.fit(X, dimensions = dim)

## Evaluation
val <- list(x = -1.5, y = 3)
eval.motbf(P, values = val)
val <- list(x = -1.5)
eval.motbf(P, values = val)
val <- list(y = 3)
eval.motbf(P, values = val)

##############################################################################
## MORE EXAMPLES #############################################################
############################################################################## 
\donttest{
## Dataset with 3 variables
X <- data.frame(x = rnorm(100), y = rexp(100), z = rnorm(100, 1))

## Joint function
dim <- c(2,1,3)
P <- jointmotbf.fit(X, dimensions = dim)
P

## Evaluation
val <- list(x = 0.8, y = -2.1, z = 1.2)
eval.motbf(P, values = val)
val <- list(x = 0.8, y = 1.2)
eval.motbf(P, values = val)
val <- list(y = -2.1)
eval.motbf(P, values = val)
}
}
