% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{evalJointFunction}
\alias{evalJointFunction}
\title{Evaluation of joint MoTBFs}
\usage{
evalJointFunction(P, values)
}
\arguments{
\item{P}{A \code{"jointmotbf"} object.}

\item{values}{A list with the name of the variables equal to the values to be evaluated.}
}
\value{
If all the variables in the equation are evaluated then a \code{"numeric"} value
is returned. Otherwise, an \code{"motbf"} object or a \code{"jointmotbf"} object is returned.
}
\description{
Evaluates a \code{"jointmotbf"} object at a specific point.
This function is deprecated; use \code{"eval.mop"} instead.
}
