% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mop.R
\name{expectedValueMOP}
\alias{expectedValueMOP}
\title{Expected Value of an MoP Density Function}
\usage{
expectedValueMOP(fx)
}
\arguments{
\item{fx}{An object of class \code{'mop'} representing an MoP probability density function.}
}
\value{
A numeric value representing the expected value of the distribution.
}
\description{
Computes the expected value (mean) of a Mixture of Polynomials (MoP) function over its domain.
}
