% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mte.R
\name{expectedValueMTE}
\alias{expectedValueMTE}
\title{Expected Value of an MTE Density Function}
\usage{
expectedValueMTE(fx)
}
\arguments{
\item{fx}{An object of class \code{'mte'} representing an MTE probability density function.}
}
\value{
A numeric value representing the expected value of the distribution.
}
\description{
Computes the expected value (mean) of a Mixture of Truncated Exponentials (MTE) function over its domain.
}
