% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inference.R
\name{get_approx_posterior}
\alias{get_approx_posterior}
\title{Approximate inference}
\usage{
get_approx_posterior(
  bn,
  target,
  evidence = NULL,
  size = 100,
  parallel = FALSE,
  ...
)
}
\arguments{
\item{bn}{An object of class \code{motbf_fit}, obtained from function \link{motbf.fit}.}

\item{target}{A character string equal to the name of the variable of interest.}

\item{evidence}{A \code{data.frame} of one row containing the value of the observed variables.}

\item{size}{A non-negative integer giving the number of random samples to generate from \code{bn}.}

\item{parallel}{\code{logical} indicating if the particle generation should be parallelized. As a default, it is set to FALSE.}

\item{...}{Optional arguments passed on to the \code{\link{univMoTBF}} function. \code{evalRange}, \code{nparam} and \code{maxParam} can be specified. \code{POTENTIAL_TYPE} is taken from the 'bn' object.}
}
\value{
A list of two elements: 1) the posterior probability distribution of the target variable, and
2) a data.frame with the generated sample, whose weights are attached as an attribute called weights (if \code{evidence} is not NULL).
}
\description{
\code{get_approx_posterior()} returns an approximation to the posterior probability distribution
of a target variable given a set of observed variables. The inference process is based on sample generation. See details.
}
\details{
If any node is observed, i.e., argument \code{evidence} is not NULL,
samples are generated from the Bayesian network using the likelihood weighting algorithm.
Otherwise, i.e., no node is observed, samples are generated using the forward sampling algorithm.
}
\examples{

## Dataset
  data("ecoli", package = "MoTBFs")
  data <- ecoli[,-c(1,9)]

## Get directed acyclic graph
  dag <- LearningHC(data)
  
## Learn bayesian network
  bn <- motbf.fit(dag, data = data, numIntervals = 4, POTENTIAL_TYPE = "MOP")
  
## Specify the evidence set and target variable
  obs <- data.frame(lip = "0.48", alm1 = 0.55, gvh = 1, stringsAsFactors=FALSE)
  node <- "alm2"
  
## Get the posterior distribution of 'node' given "evidence" and the generated sample
  get_approx_posterior(bn, target = node, evidence = obs, size = 10, maxParam = 15)
  
}
\references{
Henrion, M. (1988). Propagating uncertainty in Bayesian networks by probabilistic logic sampling. In Machine Intelligence and Pattern Recognition (Vol. 5, pp. 149-163). North-Holland.
}
