% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearningBN.R
\name{getDAG}
\alias{getDAG}
\title{Retrieve DAG from BN}
\usage{
getDAG(bn)
}
\arguments{
\item{bn}{An object of class \code{motbf_fit} or \code{bn.fit} (from bnlearn package).}
}
\value{
An object of class \code{bn} (same class as in bnlearn package).
}
\description{
Get the Directed acyclic graph (DAG) from a Bayesian network (BN)
}
