% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint.R
\name{getMotbfDim}
\alias{getMotbfDim}
\title{Extract Dimension of MoTBFs}
\usage{
getMotbfDim(P)
}
\arguments{
\item{P}{An object of class \code{"motbf"} and subclass 'mop' or \code{"jointmotbf"}.}
}
\value{
Dimension of the function.
}
\description{
Get the dimension of \code{"motbf"} and \code{"jointmotbf"} densities.
}
\examples{
## 1. EXAMPLE 
## Data
X <- rnorm(2000)

## Univariate function
f <- univMoTBF(X, POTENTIAL_TYPE = "MOP")
getMotbfDim(f)

## 2. EXAMPLE 
## Dataset with 2 variables
X <- data.frame(x = rnorm(100), y = rnorm(100))

## Joint function
dim <- c(2,3)
P <- jointmotbf.fit(X, dimensions = dim)

## Dimension of the joint function
getMotbfDim(P)

}
\seealso{
\link{univMoTBF} and \link{jointmotbf.fit}
}
