% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint.R
\name{getMotbfVar}
\alias{getMotbfVar}
\title{Extract Variables of MoTBFs}
\usage{
getMotbfVar(P)
}
\arguments{
\item{P}{An object of class \code{"univmotbf"} or \code{"jointmotbf"}.}
}
\value{
A \code{"character"} vector with the names of the variables in the function.
}
\description{
Get the names of the variables involved in a  \code{"univmotbf"} or \code{jointmotbf} object.
}
\examples{

# 1. EXAMPLE
## Generate a dataset
data <- data.frame(X1 = rnorm(100), X2 = rnorm(100))

## Joint function
dim <-c(3,2)
P <- jointmotbf.fit(data, dimensions = dim)
P

## Variables
getMotbfVar(P)

##############################################################################
## MORE EXAMPLES #############################################################
##############################################################################
\donttest{
## Generate a dataset
data <- data.frame(X1 = rnorm(100), X2 = rnorm(100), X3 = rnorm(100))

## Joint function
dim <- c(2,1,3)
P <- jointmotbf.fit(data, dimensions = dim)

## Variables
getMotbfVar(P)
}
}
