% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorKnowledge.R
\name{getNonNormalisedRandomMoTBF}
\alias{getNonNormalisedRandomMoTBF}
\title{Ramdom MoTBF}
\usage{
getNonNormalisedRandomMoTBF(degree, POTENTIAL_TYPE = "MOP")
}
\arguments{
\item{degree}{A \code{"numeric"} value containing the degree of the random function.}

\item{POTENTIAL_TYPE}{A \code{"character"} string specifying the posibles potential
types, must be one of \code{"MOP"} or \code{"MTE"}.}
}
\value{
A \code{"numeric"} vector of length 2 giving the coefficients.
}
\description{
Generates a non normalized (i.e. not integrating to 1) positive MoTBF function.
}
\examples{

getNonNormalisedRandomMoTBF(8, POTENTIAL_TYPE = "MOP")
getNonNormalisedRandomMoTBF(11, POTENTIAL_TYPE = "MTE")

}
