% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structuralLearning.R
\name{getStructure}
\alias{getStructure}
\title{Hybrid Bayesian Network structure learning}
\usage{
getStructure(data, method, target = NULL)
}
\arguments{
\item{data}{A dataset with discrete and continuous variables. If the discrete
variables are not of class \code{"factor"}, they are automatically converted.}

\item{method}{A \code{"character"} string indicating the method to learn the structure:
NB (naive Bayes), TAN (Tree augmented Naive Bayes), or HC (hill climbing) are the available options.}

\item{target}{An optional parameter only used in the case of NB and TAN to specify the class variable.}
}
\value{
The output is a \code{"bn"} object containing the learned graph.
}
\description{
Learn the structure of a hybrid Bayesian network, using a fixed method
(Naive Bayes, NB), a restricted method (Tree augmented Naive Bayes, TAN),
or an unrestricted method (the \bold{hill climbing}, HC, score-based local search method).
}
\details{
\code{getStructure()} automatically converts non-numeric variables into factors
before calling function \code{hc()} from the \code{bnlearn} package. In the case of TAN, it converts
all numeric and non-numeric variables into factors (using 4 equal width intervals)
before calling \code{tree.bayes()} from the \code{bnlearn} package.
}
\examples{

## Data
data(ecoli)
ecoli <- ecoli[,-1] ## Sequence Name

## DAG1
dag1 <- getStructure(ecoli, method = "HC")
dag1
plot(dag1)

## DAG2
dag2 <- getStructure(ecoli, method = "TAN", target = "mcg")
dag2
plot(dag2)


}
\seealso{
\link[bnlearn]{hc}
}
