% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{integralMOP}
\alias{integralMOP}
\title{Integration of MOPs}
\usage{
integralMOP(fx)
}
\arguments{
\item{fx}{An \code{"motbf"} object of subclass \code{'mop'}.}
}
\value{
The non-defined integral of the function.
}
\description{
Method to calculate the non-defined integral of an \code{"motbf"} object of \code{'mop'} subclass.
This function is deprecated; use \code{"integrate.motbf"} instead.
}
\seealso{
\link{univMoTBF} for learning and \link{integralMoTBF}
for a more complete function to get defined and non-defined integrals
of class \code{"motbf"}.
}
