% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{integralMoTBF}
\alias{integralMoTBF}
\title{Integrating MoTBFs}
\usage{
integralMoTBF(fx, min = NULL, max = NULL)
}
\arguments{
\item{fx}{An object of class \code{"motbf"}.}

\item{min}{The lower integration limit. By default it is NULL.}

\item{max}{The upper integration limit. By default it is NULL.}
}
\value{
\code{integralMoTBF()} returns either the indefinite integral of the MoTBF
function, which is also an object of class \code{"motbf"}, or the definite integral,
wich is a \code{"numeric"} value.
}
\description{
Compute the integral of a one-dimensional mixture of truncated basis function
over a bounded or unbounded interval.
This function is deprecated; use \code{"integrate.motbf"} instead.
}
\details{
If the limits of the interval, min and max are NULL, then the output is
the expression of the indefinite integral. If only 'min' contains a numeric value,
then the expression of the integral is evaluated at this point.
}
\seealso{
\link{univMoTBF}, \link{integralMOP} and \link{integralMTE}
}
