% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{integralMTE}
\alias{integralMTE}
\title{Integrating MTEs}
\usage{
integralMTE(fx)
}
\arguments{
\item{fx}{An \code{"motbf"} object of subclass \code{'mte'}.}
}
\value{
The non-defined integral of the function.
}
\description{
Method to calculate the non-defined integral of an \code{"motbf"} object of \code{'mte'}
subclass. This function is deprecated; use \code{"integrate.motbf"} instead.
}
\seealso{
\link{univMoTBF} for learning and \link{integralMoTBF}
for a more complete function to get defined and non-defined integrals
of class \code{"motbf"}.
}
