% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoTBFClass.R
\name{is.motbf}
\alias{is.motbf}
\alias{is.univmotbf}
\alias{is.jointmotbf}
\alias{is.motbf_fit}
\alias{is.motbf_fit_cv}
\alias{is.mte}
\alias{is.mop}
\alias{subclass}
\title{Check MoTBF Classes and Subclasses}
\usage{
is.motbf(x, class = "motbf")

is.univmotbf(x, class = "univmotbf")

is.jointmotbf(x, class = "jointmotbf")

is.motbf_fit(x, class = "motbf_fit")

is.motbf_fit_cv(x, class = "motbf_fit_cv")

is.mte(x)

is.mop(x)

subclass(fx)
}
\arguments{
\item{x}{An object to be checked.}

\item{class}{Character string specifying the target class name.}

\item{fx}{An object to determine the subclass for.}
}
\value{
\itemize{
\item \code{is.*}: Logical value (\code{TRUE} or \code{FALSE}) indicating if the object belongs to the checked class.
\item \code{subclass}: A character string (\code{"mop"} or \code{"mte"}) specifying the underlying family of the object.
}
}
\description{
Utility functions to check whether an object belongs to a specific MoTBF class, or to identify its underlying subclass (\code{'mop'} or \code{'mte'}).
}
