% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint.R
\name{jointmotbf.fit}
\alias{jointmotbf.fit}
\title{Joint MoTBF density learning}
\usage{
jointmotbf.fit(
  X,
  ranges = NULL,
  dimensions = NULL,
  fitPoints = 10,
  constraints = 10
)
}
\arguments{
\item{X}{a dataset of class \code{"data.frame"}.}

\item{ranges}{a \code{"numeric"} matrix containing the range of the variables used to fit the function,
where each column corresponds to a variable. If not specified, the range of each variable is computed from the data.}

\item{dimensions}{a \code{"numeric"} vector containing the number of parameters of each variable.}

\item{fitPoints}{an \code{"integer"} indicating the number of points per variable to use to build the expanded grid
where the objective function will be evaluated when optimizing the parameters.}

\item{constraints}{an \code{"integer"} indicating the number of constraints under which to minimize the quadratic function.}
}
\value{
\code{jointmotbf.fit()} returns a list with the following elements:
\item{\code{Function}}{The analytical expression of the learned density.}
\item{\code{Domain}}{A \code{"matrix"} containing the domain of each variable over which the density is defined.}
\item{\code{Iterations}}{The number of iterations needed to solve the problem.}
\item{\code{Time}}{The execution time.}
}
\description{
Function for learning joint MoTBFs.
The \code{jointmotbf.fit()} function is a wrapper of two internal (non-exported) functions:
\code{getParamJoint()} and \code{fixParamJoint()}.
The first one gets the parameters by solving a quadratic optimization problem, minimizing
the mean squared error between the empirical joint CDF and the estimated CDF.
The density is obtained as the derivative of the estimated CDF.
The second one, \code{fixParamJoint()}, fixes the equation of the joint function using
the previously learned parameters and converting this \code{"character"} string into an
object of class \code{"jointmotbf"}.
}
\examples{
## 1. EXAMPLE 
## Generate a multinormal dataset
data <- data.frame(X1 = rnorm(100), X2 = rnorm(100))

## Joint learnings
dim <- c(2,3)
P <- jointmotbf.fit(data, dimensions = dim)

P
attributes(P)
class(P)

###############################################################################
## MORE EXAMPLES ##############################################################
###############################################################################
\donttest{
## Generate a dataset
data <- data.frame(X1 = rnorm(100), X2 = rnorm(100), X3 = rnorm(100))

## Joint learnings
dim <- c(3,2,3)
P <- jointmotbf.fit(data, dimensions = dim)
P
attributes(P)
class(P)
}
}
