% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint.R
\name{marginal.jointmotbf}
\alias{marginal.jointmotbf}
\title{Marginalization of MoTBFs}
\usage{
marginal.jointmotbf(P, var)
}
\arguments{
\item{P}{An object of class \code{"jointmotbf"}, i.e., the joint density function.}

\item{var}{A vector containing the names of the marginal variables (those being retained in 'P').
This argument accepts the \code{"numeric"} position (w.r.t. P$Domain) or the \code{"character"} name of the variables.}
}
\value{
If 'var' is an atomic vector, the marginal distribution of the variable specified in 'var' is computed
from the joint distribution and the result is an object of class \code{"univmotbf"}, i.e., a univariate distribution.

If 'var' contains more than one variable, the joint distribution over this subset is computed, i.e.,
the variables NOT contained in 'var' are marginalized out and the result is an object of class \code{"jointmotbf"}.
}
\description{
Computes the marginal densities from a \code{"jointmotbf"} object.
}
\examples{
## 1. EXAMPLE 
## Dataset with 2 variables
data <- data.frame(x = rnorm(100), y = rnorm(100))

## Joint function
dim <- c(4,3)
P <- jointmotbf.fit(data, dimensions = dim)

## Marginal
marginal.jointmotbf(P, var = "x")
marginal.jointmotbf(P, var = 2)

##############################################################################
## MORE EXAMPLES #############################################################
##############################################################################
\donttest{
## Generate a dataset with 3 variables
data <- data.frame(x = rnorm(100), y = rnorm(100), z = rnorm(100))

## Joint function
dim <- c(2,2,3)
P <- jointmotbf.fit(data, dimensions = dim)

## Marginal
marginal.jointmotbf(P, var = "x")
marginal.jointmotbf(P, var = "y")
marginal.jointmotbf(P, var = c("x", "z"))
}
}
\seealso{
\link{jointmotbf.fit} and \link{eval.motbf}
}
