% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{marginalJointMoTBF}
\alias{marginalJointMoTBF}
\title{Marginalization of MoTBFs}
\usage{
marginalJointMoTBF(P, var)
}
\arguments{
\item{P}{An object of class \code{"jointmotbf"}, i.e., the joint density function.}

\item{var}{The \code{"numeric"} position or the \code{"character"} name of the marginal variable.}
}
\value{
The marginal of a \code{"jointmotbf"} function. The result is an object of class \code{"motbf"}.
}
\description{
Computes the marginal densities from a \code{"jointmotbf"}
object.
This function is deprecated; use \code{"marginal.jointmotbf"} instead.
}
\seealso{
\link{jointMoTBF} and \link{evalJointFunction}
}
