% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI_MoP_TAN.R
\name{MOPTAN}
\alias{MOPTAN}
\alias{fit_tan}
\alias{mutual_information_tan}
\title{Fitting MoTBFs TAN models}
\usage{
fit_tan(
  target,
  data,
  fit.args = NULL,
  root = NULL,
  all = FALSE,
  mutualInfoCond = NULL,
  parallel = FALSE
)

mutual_information_tan(data, target, fit.args = NULL, parallel = FALSE)
}
\arguments{
\item{target}{A \code{character} string indicating the name of the target or
class variable. Target must be a column of \code{data} and it can be a
continuous or discrete variable.}

\item{data}{A \code{data.frame} containing the variables, which can contain
continuous and discrete variables.}

\item{fit.args}{A \code{list} a list containing optional arguments used to
fit the models.
These arguments must be those accepted by function \link{motbf.fit},
i.e., 'numIntervals' (4), 'POTENTIAL_TYPE' ('MOP'), 'maxParam' (7),
's' (NULL), 'priorData' (NULL) or 'scale' (TRUE).
If fit.args is left NULL, the default values (in brackets) for those
arguments will be used.}

\item{root}{A \code{character} string indicating the label of the root
predictor variable of TAN model.}

\item{all}{A \code{logical} flag. If \code{TRUE}, the function return a
\code{list} which contains two elements: the TAN model and the mutual
information used to compute the maximum spanning tree.
Defaults to \code{FALSE}.}

\item{mutualInfoCond}{A numeric matrix indicating the estimation of the
mutual information coefficientes for Chow-Liu- algorithm in TAN. If it is
\code{NULL}, it is computed using \link{mutual_information_tan} function.}

\item{parallel}{A \code{logical} flag. If \code{TRUE}, computation runs in parallel
using \code{foreach} and \code{doParallel}. Defaults to \code{FALSE}.}
}
\value{
The main function, \code{fit_tan()}, returns an object of class \code{"motbf_fit"}.
When \code{all=TRUE}, it returns a list with the Bayesian network and the mutual
information matrix used to compute the maximun spanning tree.

Function \code{mutual_information_tan()} returns a symmetric numeric
\code{matrix} of dimensions \eqn{k \times k}{k x k}, where \eqn{k} is the number of
predictor variables. Row and column names correspond to the predictor
variables, and the entries contain the estimated conditional mutual
information values. This matrix is used to compute the TAN model.
}
\description{
Perform a TAN model of class MoTBF based on maximizing the Mutual Information.
}
\details{
The main function, \code{fit_tan()}, fits a MoTBF Tree Augmented Naive Bayes
model using the specified \code{data}.
}
\examples{
\donttest{
data = iris
data$Species = as.factor(data$Species)
# Fit TAN model for classification
tan = fit_tan("Species",data)
}

}
