% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidation.R
\name{motbf.cv}
\alias{motbf.cv}
\title{Cross-validation for MoTBFs}
\usage{
motbf.cv(
  data,
  dag,
  loss,
  k = 10,
  target = NULL,
  seed = NULL,
  fit.args = NULL,
  loss.args = NULL,
  foldsIndex = NULL,
  ...
)
}
\arguments{
\item{data}{an object of class \code{"data.frame"}, which can contain continuous and discrete variables.}

\item{dag}{a network of the class \code{"bn"}, \code{"graphNEL"} or \code{"network"}.}

\item{loss}{a character string indicating which loss function should be used.
Currently, two options are available: 'logl', for the log-likelihood of the model; and 'pred', for the predictive error. See details.}

\item{k}{an integer indicating the number of folds to split the data set. If k =  0, the train and test sets
are the same data; if k = 1, hold-out validation is carried out, i.e., the data set is split in train (80\% by default) and test (20\% by default);
finally, if k >=2, k-fold cross validation is carried out.}

\item{target}{a character string indicating which node is the target. This argument might be NULL if
the loss function chosen is the log-likelihood of the model ('logl').}

\item{seed}{an integer to specify the seed.
The k-folds are created randomly, so one might expect slightly different results unless 'seed' is used.}

\item{fit.args}{a list containing optional arguments used to fit the models. These arguments must be those accepted by function
\link{motbf.fit}, i.e., 'numIntervals' (4), 'POTENTIAL_TYPE' ('MOP'), 'maxParam' (NULL), 's' (NULL), 'priorData' (NULL) or 'scale' (TRUE).
If fit.args is left NULL, the default values (in brackets) for those arguments will be used.}

\item{loss.args}{a list containing optional arguments related to the loss functions. Currently available arguments are:
'loss.matrix' and 'percentage_test'. See details.}

\item{foldsIndex}{a list containing the test indexes for each fold of cross-validation. If is not null, \code{k} and \code{seed} are ignored.}

\item{...}{Additional arguments. Not used currently.}
}
\value{
An object of class \code{"motbf.fit.cv"}. This is a list of \code{"k"} elements,
each of them containing the results of each fold. More specifically, each element contains:
\describe{
\item{test}{a \code{"data.frame"} of the subset used to fit the model.}
\item{fitted}{an object of class \code{"motbf.fit"}, i.e., the model fitted.}
\item{loss}{the loss value computed for the fold.
If \code{"pred"} is chosen for the argument \code{"loss"}, then each element of the \code{"motbf.fit.cv"} object
also contains:
\describe{
\item{predicted}{a vector of predictions for the target variable.}
\item{observed}{a vector of observed records of the target variable.}
}
}
}
}
\description{
Perform a k-fold cross validation for Bayesian networks of class MoTBF.
}
\details{
If the basis function is MTE, the loss function is the log-likelihood of the model.
On the other hand, if the basis function if MOP, the loss function might be either
the predictive error ('pred') or the the log-likelihood of the model ('logl').

Details on the \code{loss} argument:
\describe{
\item{\code{'logl'}}{The log-likelihood of the model is computed. This measure is available for both basis functions, MTE and MOP.}
\item{\code{'pred'}}{This option is only available for MOPs.
The predictive error (root mean squared error) or classification accuracy is computed as the loss function,
depending on the nature of the target variable (continuous or discrete, respectively).
The program will guess which type the target variable is and compute the corresponding measure.}
}

Details on the \code{loss.args} argument. Currently, two arguments can be specified within this list:
\describe{
\item{\code{'p.test'}}{Only used if \code{k = 1}.
This argument specifies the proportion of the data set that goes to the test set (between 0 and 1).}
\item{\code{'loss.matrix'}}{A squared \code{matrix} used to compute a weighted classification accuracy for discrete targets.
This matrix is multiplied by the confusion matrix element by element,
and the resulting matrix is used to compute the classification accuracy.
The loss matrix allows to increase the penalty of some user-specified errors.
If the loss matrix provided is a constant matrix of ones, the result is the standard classification accuracy.
Note that the diagonal of the loss matrix is regarded as a reward, while the off-diagonal is regarded as a cost.
}
}
}
\examples{
#################
### EXAMPLE 1 ###
#################
## Perform 2-fold cross validation using the default model arguments
## and the log-likelihood as loss function

# Load data
data(ecoli)
ecoli <- ecoli[,-c(1,9)]

# Learn DAG
dag <- LearningHC(ecoli)

# Run cross validation
cv = motbf.cv(data = ecoli, dag, k = 2, loss = 'logl')
cv

\donttest{
#################
### EXAMPLE 2 ###
#################
## Choose different arguments to fit the model parameters
fit.args = list(numIntervals = 3, POTENTIAL_TYPE = 'MOP', maxParam = 4)

# Run cross validation using the classification accuracy as loss function
cv = motbf.cv(data = ecoli, dag, k = 2, loss = 'pred', target = 'lip', 
  fit.args = fit.args)
cv
summary(cv)

#################
### EXAMPLE 3 ###
#################

## Specify a loss matrix to increase the penalty of classification errors
lossFunctionMatrix = matrix(c(c(1,2), c(3, 1)),nrow = 2, ncol = 2, byrow = TRUE)

# Run cross validation using the weighted classification accuracy as loss function
cv = motbf.cv(data = ecoli, dag, k = 2, loss = 'pred', target = 'lip', 
  loss.args = list(loss.matrix = lossFunctionMatrix))
cv
summary(cv)
  
#################
### EXAMPLE 4 ###
#################

# Run cross validation using the root mean squared error as loss function
cv = motbf.cv(data = ecoli, dag, k = 2, loss = 'pred', target = 'mcg')
cv
}
}
