% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableEliminationFunctionsClean.R
\name{motbf2bnlearn}
\alias{motbf2bnlearn}
\title{Export discrete motbf to bnlearn format}
\usage{
motbf2bnlearn(bn)
}
\arguments{
\item{bn}{An object of class \code{motbf_fit}, obtained from function \link{motbf.fit}. Only discrete networks are valid.}
}
\value{
An object of class \code{bn}.
}
\description{
Export a discrete BN created with \link{motbf.fit} to an object of class \code{"bn"} of package bnlearn.
}
