% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableEliminationFunctionsClean.R
\name{motbf2grain}
\alias{motbf2grain}
\title{Export discrete motbf to grain format}
\usage{
motbf2grain(bn)
}
\arguments{
\item{bn}{An object of class \code{motbf_fit}, obtained from function \link{motbf.fit}. Only discrete networks are valid.}
}
\value{
An object of class \code{grain}.
}
\description{
Export a discrete BN created with \link{motbf.fit} to an object of class \code{"grain"} of package gRain
}
