% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{nVariables}
\alias{nVariables}
\title{Number of Variables in a Joint Function}
\usage{
nVariables(P)
}
\arguments{
\item{P}{An \code{"motbf"} object or a \code{"jointmotbf"} object.}
}
\value{
A \code{"character"} vector with the names of the variables in the function.
}
\description{
Compute the number of variables which are in a \code{jointmotbf} object.
This function is deprecated; use \code{"getMotbfVar"} instead.
}
