% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.motbf}
\alias{plot.motbf}
\alias{plot.univmotbf}
\alias{plot.piecewisemop}
\alias{plot.motbf.fit.node}
\alias{plot.jointmotbf}
\title{Plots for \code{'motbf'} objects}
\usage{
\method{plot}{motbf}(x, ...)

\method{plot}{univmotbf}(x, xlim = NULL, ylim = NULL, type = "l", add = FALSE, ...)

\method{plot}{piecewisemop}(x, xlim = NULL, ylim = NULL, ...)

\method{plot}{motbf.fit.node}(x, panels = TRUE, ...)

\method{plot}{jointmotbf}(
  x,
  type = "contour",
  ranges = NULL,
  orientation = c(5, -30),
  data = NULL,
  filled = TRUE,
  ticktype = "simple",
  main = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class \code{'motbf'} or one of its subclasses (\code{'univmotbf'}, \code{'piecewisemop'}, \code{'motbf.fit.node'}, or \code{'jointmotbf'}).}

\item{\dots}{Further arguments to be passed as for \link{plot}.}

\item{xlim, ylim}{Numeric vectors of length 2 specifying x and y axis limits. By default \code{0:1}. Used in \code{'univmotbf'} and \code{'piecewisemop'}.}

\item{type}{Character string specifying the plot type, as for \link{plot}. For \code{'univmotbf'} defaults to \code{'l'} (line); for \code{'jointmotbf'} defaults to \code{"contour"} (another option is \code{"perspective"}).}

\item{add}{Logical; if \code{TRUE}, adds the plot to the existing active graphics device. Used only in \code{'univmotbf'}.}

\item{panels}{Logical; if \code{TRUE}, displays plots in separate panels. Used only in \code{'motbf.fit.node'}.}

\item{ranges}{A \code{"numeric"} matrix containing the domain of the variables, by columns, which is used to specify the plotting range. Used only in \code{'jointmotbf'}.}

\item{orientation}{A \code{"numeric"} vector indicating the perpective of the plot in degrees. By default, it is set to \code{(5,-30)}. Used only in \code{'jointmotbf'}.}

\item{data}{An object of class \code{"data.frame"} containing two columns only.
This argument is used to draw the points over the main plot. By default, it is set to \code{NULL}. Used only in \code{'jointmotbf'}.}

\item{filled}{A logical argument; it is only used if \code{type = "contour"}.
is active. By default, it is \code{TRUE}, so filled contours are plotted. Used only in \code{'jointmotbf'}.}

\item{ticktype}{A \code{"character"} string, either \emph{simple} or \emph{detailed}. By default, it is set to \code{"simple"},
which draws just an arrow parallel to the axis to indicate direction of increase.
In contrast, \code{"detailed"} draws normal ticks. This argument is only used in the \code{"perspective"} plot. Used only in \code{'jointmotbf'}.}

\item{main}{Title string for the plot. Used in \code{'jointmotbf'}.}
}
\value{
A plot of the specificated function.
}
\description{
Draws an \code{'motbf'} function.
}
\examples{

## 1. univmotbf - Example for univariate distributions 
## Data
X <- rexp(2000)

f1 <- univMoTBF(X, POTENTIAL_TYPE = "MOP")
f2 <- univMoTBF(X, POTENTIAL_TYPE = "MTE", maxParam = 10)

## Plots
plot(NULL, xlim = range(X), ylim = c(0,0.8), xlab="X", ylab="density")
plot(f1, xlim = range(X), col = 1, add = TRUE)
plot(f2, xlim = range(X), col = 2, add = TRUE)


## 2. jointmotbf - Example for join distributions 
## Data
X <- data.frame(rnorm(500), rnorm(500))

## Joint function
dim <- c(3,3) 
P <- jointmotbf.fit(X, dimensions = dim)

## Plots
plot(P)
plot(P, type = "perspective", orientation = c(90,0))

}
