% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.motbf_fit}
\alias{predict.motbf_fit}
\title{Predict from an MoTBF Bayesian Network}
\usage{
\method{predict}{motbf_fit}(
  object,
  target,
  data,
  method = NULL,
  prob = FALSE,
  parallel = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of class "motbf_fit"}

\item{target}{A character string indicating the variable to be predicted}

\item{data}{A data frame containing the predictive variables}

\item{method}{A character string indicating the method to perform the inference. Options are NULL, "ve" (variable elimination) and "fs" (forward sampling).}

\item{prob}{A logical value. If TRUE, the probabilities used for classification are returned as an attribute.}

\item{parallel}{A logical value. If TRUE, parallelization is carried out.}

\item{...}{Additional arguments used if method = "fs".}
}
\description{
Predict from an MoTBF Bayesian Network
}
