% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.motbf}
\alias{print.motbf}
\alias{print.motbf_fit}
\alias{print.motbf.fit.node}
\alias{print.motbf_fit_cv}
\alias{print.univmotbf}
\alias{print.piecewisemop}
\alias{print.jointmotbf}
\title{Print object of class motbf
\code{print} method for class \code{"motbf"}.}
\usage{
\method{print}{motbf}(x, ...)

\method{print}{motbf_fit}(x, ...)

\method{print}{motbf.fit.node}(x, ...)

\method{print}{motbf_fit_cv}(x, ...)

\method{print}{univmotbf}(x, ...)

\method{print}{piecewisemop}(x, ...)

\method{print}{jointmotbf}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"motbf"}, \code{"motbf_fit"}, \code{"univmotbf"}, \code{"piecewisemop"}, \code{"jointmotbf"}, or \code{"motbf.fit.cv"}.}

\item{...}{optional arguments passed to print for other classes created in the MoTBFs package.
Currently, no optional arguments are supported.}
}
\description{
Print object of class motbf
\code{print} method for class \code{"motbf"}.

Print a single node of a BN.
This function is called by print.motbf_fit, but not exported.

Print the results of a k-fold cross validation
}
\details{
The following classes are created in the MoTBFs package:
\describe{
\item{motbf}{the generic class common to all objects created in the package}
\item{univmotbf}{the class corresponding to MTE or MOP univariate distributions. An object of class \code{univmotbf} is the output of function \code{univMoTBF()}.}
\item{piecewisemop}{the class corresponding to MOP univariate distributions defined by multiple sub-functions with different domain. When calling \code{variableElimination()}, the output might be of this class.}
\item{jointmotbf}{the class corresponding to joint distributions. An object of class \code{jointmotbf} is the output of function \code{jointMOP()}}
\item{motbf_fit}{the class corresponding to fully fitted Bayesian network models (either discrete, continuous or hybrid). An object of class \code{motbf_fit} is the output of function \code{motbf.fit()}.}
\item{motbf.fit.cv}{the class corresponding to k-fold cross validation results. An object of class \code{motbf.fit.cv} is the output of function \code{motbf.cv()}.}
\item{motbf.fit.node}{the class corresponding to a single node of a Bayesian Network.}
}
}
\examples{
\donttest{
## Dataset Ecoli
data(ecoli)
data <- ecoli[,-c(1)] ## remove variable sequence

## Directed acyclic graph
dag <- LearningHC(data)

## Learning BN
P <- motbf.fit(graph = dag, data = data, numIntervals = 3, POTENTIAL_TYPE = "MOP",
maxParam = 15)
P
}

}
