% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteLearning.R
\name{printDiscreteBN}
\alias{printDiscreteBN}
\title{Printing discrete Bayesian networks}
\usage{
printDiscreteBN(BN)
}
\arguments{
\item{BN}{A discrete learning.}
}
\value{
The results are shown on the screen.
}
\description{
Prints the univariate and conditional distributions of a discrete BN.
}
