% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteLearning.R
\name{probDiscreteVariable}
\alias{probDiscreteVariable}
\title{Probability distribution of discrete variables}
\usage{
probDiscreteVariable(x)
}
\arguments{
\item{x}{a \code{"factor"} containing the records of the discrete variable.}
}
\value{
A list of 2 elements:
\item{coeff}{a named vector that contains the probabilities.}
\item{sizeDataLeaf}{a vector containing the number of records in each leaf of the discrete tree.}
}
\description{
Compute the probabilities of a discrete variable from a data set using Laplace correction.
}
\details{
Laplace correction is used to avoid the zero probability problem, with \eqn{\alpha = 1} as smoothing factor.
Therefore, the probability for each state, \eqn{\theta_i}, is computed as
\deqn{\theta_i = \frac{x_i + 1}{N + K}}{%
 \theta_i =  (x_i + 1)/(N + K)}
for \eqn{i = 0, 1, \ldots, K},
where \eqn{x_i} is the number of records of state \eqn{i},
\eqn{N} is the total number of records, and
\eqn{K} is the total number of states.
}
\examples{
## Simulate discrete variable
x <- factor(sample(c('yes', 'no', 'maybe'), 500, replace = TRUE), 
  levels = c('yes', 'no', 'maybe'))
 
## Compute probabilities
p <- probDiscreteVariable(x)
p


}
\seealso{
\link{discreteVariablesStates}
}
