% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inference.R
\name{query}
\alias{query}
\title{Conditional probability queries}
\usage{
query(sample, event, evidence)
}
\arguments{
\item{sample}{A \code{data.frame} containing a sample.}

\item{event}{A expression describing the event of interest}

\item{evidence}{A expression describing the conditioning evidence.}
}
\value{
The conditional probability: P(event|evidence).
}
\description{
Compute conditional probability queries from a sample.
}
\examples{
  data("ecoli", package = "MoTBFs")
  dat <- ecoli[,-c(1,4,5,9)]
  
  # Build DAG
  dag <- LearningHC(dat)
  
  # Learn BN parameters
  bn = motbf.fit(dag, dat)
 
 # Get sample from bn
 sam = sample_motbfs(bn, 50)
 
 # Compute P(mcg > 0.6 | aac < 0.8 & alm2 < 0.3)
 query(sam, event = (mcg >0.6), evidence = (aac<0.8 & alm2 <0.3))
 
}
