% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescalatedFunctions.R
\name{rescale_data}
\alias{rescale_data}
\title{Scale data}
\usage{
rescale_data(data, v_mean = NULL, v_sd = NULL)
}
\arguments{
\item{data}{data.frame containing the variables to be standardized.}

\item{v_mean}{vector of means of the data.frame. The value for the non-numeric columns should be NA. If not given, it is computed from data.}

\item{v_sd}{vector of standard deviations of the data.frame. The value for the non-numeric columns should be NA. If not given, it is computed from data.}
}
\value{
A data.frame with scaled data. Non-numeric columns are also returned in their original position.
}
\description{
Standardize the numeric columns of a data.frame.
}
