% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescalatedFunctions.R
\name{rescaledFunctions}
\alias{rescaledFunctions}
\alias{rescaledMoTBFs}
\alias{rescaledMOP}
\alias{rescaledMTE}
\alias{ToStringRe_MTE}
\alias{meanMOP}
\alias{rescale_motbf_fit}
\title{Rescaling MoTBF functions}
\usage{
rescaledMoTBFs(fx, data = NULL)

rescaledMOP(fx, data = NULL)

rescaledMTE(fx, data)

ToStringRe_MTE(parameters, data, num = 1)

meanMOP(fx)

rescale_motbf_fit(object, POTENTIAL_TYPE, data)
}
\arguments{
\item{fx}{A function of class \code{"motbf"} learned from a scaled data.}

\item{data}{the original data set (non-scaled)}

\item{parameters}{A \code{"numeric"} vector with the coefficients to create the rescaled MoTBF.}

\item{num}{A \code{"numeric"} value which contains the denominator of the coefficient
in the exponential. By default it is 5.}

\item{object}{and object of class motbf_fit that has been learned with scaled data}

\item{POTENTIAL_TYPE}{the potential of the model (either "MOP" or "MTE")}
}
\value{
An \code{"motbf"} function of the original data.
}
\description{
A collection of function to reescale an MoTBF function
to the original offset and scale. This is useful when data was
standardized previously to learning.
}
\examples{
## 1. EXAMPLE
X <- rchisq(1000, df = 8) ## data
modX <- scale(X) ## scale data

## Learning
f <- univMoTBF(modX, POTENTIAL_TYPE = "MOP", nparam=10) 
plot(f, xlim = range(modX), col=2)
hist(modX, prob = TRUE, add = TRUE)

## Rescale
origF <- rescaledMoTBFs(f, X) 
plot(origF, xlim = range(X), col=2)
hist(X, prob = TRUE, add = TRUE)
expectedValueMOP(origF) 
mean(X)

## 2. EXAMPLE 
X <- rweibull(1000, shape = 20, scale= 10) ## data
modX <- as.numeric(scale(X)) ## scale data

## Learning
f <- univMoTBF(modX, POTENTIAL_TYPE = "MTE", nparam = 9) 
plot(f, xlim = range(modX), col=2, main="")
hist(modX, prob = TRUE, add = TRUE)

## Rescale
origF <- rescaledMoTBFs(f, X) 
plot(origF, xlim = range(X), col=2)
hist(X, prob = TRUE, add = TRUE)
expectedValueMTE(origF)
mean(X)
}
\seealso{
\link{univMoTBF}
}
