% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inference.R
\name{sample_motbfs}
\alias{sample_motbfs}
\title{Generate Samples From an MoTBF Bayesian network}
\usage{
sample_motbfs(bn, n, parallel = FALSE, evidence = NULL)
}
\arguments{
\item{bn}{An object of class motbf_fit, obtained from the function \link{motbf.fit}.}

\item{n}{A non-negative integer giving the number of instances to be generated.}

\item{parallel}{A \code{logical} value. If TRUE, parallelization is carried out. As a default, it is set to FALSE}

\item{evidence}{A data.frame of one row containing the values for the observed variables. As a default, it is NULL.}
}
\value{
A \code{data.frame} containing the generated sample. Is evidence is not NULL, attribute 'w' contains the weight of each sample.
}
\description{
This function generates a sample from an MoTBF Bayesian network.
}
\examples{
  data("ecoli", package = "MoTBFs")
  dat <- ecoli[,-c(1,4,5,9)]
  
  # Build DAG
  dag <- LearningHC(dat)
  
  # Learn BN parameters
  bn = motbf.fit(dag, dat)
 
 # Get sample from bn
 sam = sample_motbfs(bn, 50)
 
 
}
