% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{subsetData}
\alias{subsetData}
\alias{splitFolds}
\alias{TrainingandTestData}
\alias{splitdata}
\title{Dataset subsetting}
\usage{
splitFolds(data, k)

TrainingandTestData(data, percentage_test, discreteVariables = NULL)

splitdata(data, nameVariable, min, max)
}
\arguments{
\item{data}{A dataset of class \code{data.frame}.}

\item{k}{The number of folds for k-fold cross validation, used in splitFolds function.}

\item{percentage_test}{The proportion of data that goes to the test set (between 0 and 1), used in TrainingandTestData function.}

\item{discreteVariables}{A \code{character} vector with the name of the discrete variables in the dataset.}

\item{nameVariable}{A \code{character} vector with the name of the variable to be filtered, used in splitdata function.}

\item{min, max}{Boundary values to filter out, used in splitdata function.}
}
\value{
\code{TrainingandTestData()} returns a list of 2 elements containing the train and test datasets.
\code{splitdata()} returns a subset of observations.
}
\description{
Collection of functions for subsetting a \code{"data.frame"} by rows or columns, and
to create training and test partitions.
}
\examples{
\donttest{
## Dataset
X <- rnorm(1000)
Y <- rchisq(1000, df = 8)
Z <- rep(letters[1:10], times = 1000/10)
data <- data.frame(X = X, Y = Y, Z = Z)

## Training and Test Datasets
TT <- TrainingandTestData(data, percentage_test = 0.2)
TT$Training
TT$Test

## Subset Dataset
splitdata(data, nameVariable = "X", min = 2, max= 3)
}
}
