% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableEliminationFunctionsClean.R
\name{variableElimination}
\alias{variableElimination}
\title{Exact inference}
\usage{
variableElimination(bn, target, evidence = NULL, elimOrder = NULL)
}
\arguments{
\item{bn}{An object of class \code{motbf_fit}, obtained from function \link{motbf.fit}.}

\item{target}{A character string equal to the name of the variable of interest.}

\item{evidence}{A \code{data.frame} of one row containing the value of the observed variables. A list can also be provided.}

\item{elimOrder}{The elimination order can be manually specified as a vector containing the
names of the variables, in the desired order. If elimOrder is not specified, the topological order is computed.}
}
\value{
The posterior probability distribution of the target variable as an object of
class \code{univmotbf} or \code{piecewisemop} if \code{target} is continuous,
or a matrix if \code{target} is discrete.
}
\description{
Compute the posterior distribution of a variable of interest given some evidence.
The variable elimination algorithm is used.
}
\examples{
## Dataset
  data("ecoli", package = "MoTBFs")
  data <- ecoli[,-c(1,9)]

## Get directed acyclic graph
  dag <- LearningHC(data)
  
## Learn bayesian network
  bn <- motbf.fit(dag, data = data, numIntervals = 4, POTENTIAL_TYPE = "MOP")
  
## Specify the evidence set and target variable
  obs <- data.frame(lip = "0.48", alm1 = 0.55, stringsAsFactors=FALSE)
  node <- "alm2" 
ve = variableElimination(bn, target = node, evidence = obs)
}
