% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mop.R
\name{derivMOP}
\alias{derivMOP}
\title{Derivative MOP}
\usage{
derivMOP(fx)
}
\arguments{
\item{fx}{An \code{"motbf"} object of the \code{'mop'} subclass.}
}
\value{
The derivative which is also an \code{"motbf"} function.
}
\description{
Compute the derivative of an \code{"motbf"} object with \code{'mop'} subclass.
}
\examples{

## 1. EXAMPLE
X <- rexp(1000)
Px <- univMoTBF(X, POTENTIAL_TYPE="MOP")
derivMOP(Px)

## 2. EXAMPLE
X <- rnorm(1000)
Px <- univMoTBF(X, POTENTIAL_TYPE="MOP")
derivMOP(Px)

\dontrun{
## 3. EXAMPLE
X <- rnorm(1000)
Px <- univMoTBF(X, POTENTIAL_TYPE="MTE")
derivMOP(Px)
## Error in derivMOP(Px): fx is an 'motbf' function but not 'mop' subclass.
class(Px)
subclass(Px)
}
}
\seealso{
\link{univMoTBF} for learning and \link{derivMoTBF} for 
general \code{"motbf"} models.
}
