% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{preprocessedData}
\alias{preprocessedData}
\title{Remove Missing Values in a Dataset by Rows}
\usage{
preprocessedData(data, strangeElements)
}
\arguments{
\item{data}{A dataset of class \code{"matrix"} or \code{"data.frame"},}

\item{strangeElements}{A \code{"character"} string which contains the elementes to remove.}
}
\description{
Delete rows of a dataset wich contains anomalous values.
}

