% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteLearning.R
\name{printDiscreteBN}
\alias{printDiscreteBN}
\title{Prints Discrete Learnings}
\usage{
printDiscreteBN(BN)
}
\arguments{
\item{BN}{A discrete learning.}
}
\value{
The results are shown by screen.
}
\description{
Shows the results of univariate and conditional learning of a discrete BN.
}
